refactor GetBlockInfo RPC

+++ ckb-mercury-sdk/src/main/java/org/nervos/mercury/model/req/payload/GetBalancePayload.java
@@ -3,13 +3,12 @@ package org.nervos.mercury.model.req.payload;
 import org.nervos.mercury.model.common.AssetInfo;
 import org.nervos.mercury.model.req.item.Item;
 
-import java.math.BigInteger;
 import java.util.Set;
 
 public class GetBalancePayload {
   public Item item;
   public Set<AssetInfo> assetInfos;
-  public BigInteger tipBlockNumber;
+  public Long tipBlockNumber;
 
   protected GetBalancePayload() {
   }
