+++ ckb-mercury-sdk/src/main/java/org/nervos/mercury/model/req/payload/GetBlockInfoPayload.java
@@ -1,7 +1,7 @@
 package org.nervos.mercury.model.req.payload;
 
 public class GetBlockInfoPayload {
-  public int blockNumber;
+  public Long blockNumber;
   public byte[] blockHash;
 
   protected GetBlockInfoPayload() {
