+++ ckb-mercury-sdk/src/main/java/org/nervos/mercury/model/resp/BlockInfoResponse.java
@@ -3,10 +3,9 @@ package org.nervos.mercury.model.resp;
 import java.util.List;
 
 public class BlockInfoResponse {
-  public int blockNumber;
+  public long blockNumber;
   public byte[] blockHash;
   public byte[] parentHash;
   public long timestamp;
-
   public List<TransactionInfoResponse> transactions;
 }
