+++ ckb-mercury-sdk/src/main/java/org/nervos/mercury/model/resp/BurnInfo.java
@@ -1,6 +1,8 @@
 package org.nervos.mercury.model.resp;
 
+import java.math.BigInteger;
+
 public class BurnInfo {
   public byte[] udtHash;
-  public long amount;
+  public BigInteger amount;
 }
