+++ ckb-mercury-sdk/src/main/java/org/nervos/mercury/model/resp/TransactionInfoResponse.java
@@ -2,13 +2,12 @@ package org.nervos.mercury.model.resp;
 
 import org.nervos.mercury.model.common.Record;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public class TransactionInfoResponse {
   public byte[] txHash;
   public List<Record> records;
-  public BigInteger fee;
+  public long fee;
   public List<BurnInfo> burn;
   public long timestamp;
 }
